requirements.txt
amqp==2.4.2
anyjson==0.3.3
billiard==3.6.0.0
celery==4.3.0
Django==2.2.1
djangorestframework==3.9.3
kombu==4.5.0
pkg-resources==0.0.0
psycopg2==2.8.2
pytz==2019.1
redis==3.2.1
sqlparse==0.3.0
vine==1.3.0
